Initial work on Texture and Sampler API + more
- Added stb_image
- Added Renderer::DestroyResources() function
- Added 'CompareOp' header
- Bugfixes

+++ Engine/Source/Flux/Runtime/Platform/Windows/WindowsThread.cpp
@@ -52,7 +52,7 @@ namespace Flux {
 	uint32 WindowsThread::GetRemainingJobs()
 	{
 		EnterCriticalSection(&m_CriticalSection);
-		uint32 result = static_cast<size_t>(m_Jobs.size());
+		uint32 result = static_cast<uint32>(m_Jobs.size());
 		WakeConditionVariable(&m_ConditionVariable);
 		LeaveCriticalSection(&m_CriticalSection);
 		return result;
